Rename dockingpanebase.h -> DockingPaneBase.h
