+++ DockingPaneBase.h
