+++ DockingPaneContainer.h
@@ -22,7 +22,7 @@
 
 #include <QWidget>
 #include <QGridLayout>
-#include "dockingpanebase.h"
+#include "DockingPaneBase.h"
 #include "DockingPaneTitleWidget.h"
 #include "DockingToolButton.h"
 #include "DockingPaneGlow.h"
